LocalImprovable interface added to GeneticWorld

+++ src/com/d9nich/pathFindingAlgorithm/geneticAlgorithm/localStrategy/LocalImprovable.java
@@ -5,5 +5,5 @@ import com.d9nich.pathFindingAlgorithm.geneticAlgorithm.Liveable;
 import java.util.ArrayList;
 
 public interface LocalImprovable {
-    void improve(ArrayList<? extends Liveable> animals, int liveFit);
+    void improve(ArrayList<? extends Liveable> animals, double liveFit);
 }
